+++ tests/test.py
@@ -1,0 +1,11 @@
+import yacman
+bulkercfg = "/project/shefflab/bulker/bulker_rivi.yaml"
+bc = yacman.YacAttMap(filepath=bulkercfg)
+bc
+
+
+bc.bulker.default_crate_folder
+
+manifest = yacman.YacAttMap(filepath="/home/ns5bc/code/bulker/demo/demo_manifest.yaml")
+bc
+